Odisha Coast
The code help in download the Landsat Satellite Imagery for the Odisha Coastline. The date can change according to your needs and the band selection depends upon your problem. The path and row along the odisha coast is 
path 139 and row 45, row 46; path 140 and row 46, row 47. 
The code also calculates NDWI and MNDWI . the results are saved to the drive.